Rename ctx values
We would like to distinguish between `ctx` referring to an enclosing parameter
and `ctx` referring to a field. This is so that we can at some point drop
`ctx` as a parameter name and just use (using Context). The `ctx` would then
go to an imported def in `Contexts`. But any other use of `ctx` does does not
follow strict nesting discipline can then change meaning, so we should
get these out of the way first.